fixed all the issues and organized the files
