+++ .gitignore
@@ -1,2 +1,2 @@
 settings.json
-.gitmodules
+discord.py-self
