--- .gitmodules
@@ -1,3 +1,0 @@
-[submudule "discord.py-self"]
-    path=discord.py-self
-    url=https://github.com/dolfies/discord.py-self
